refactor make minimal subject directory, now only used internally
