+++ mne_bids/tests/test_write.py
@@ -35,8 +35,7 @@ with warnings.catch_warnings():
     import mne
 
 from mne.datasets import testing
-from mne.utils import (check_version, requires_nibabel, requires_version,
-                       requires_freesurfer)
+from mne.utils import check_version, requires_nibabel, requires_version
 from mne.io import anonymize_info
 from mne.io.constants import FIFF
 from mne.io.kit.kit import get_kit_info
@@ -44,8 +43,7 @@ from mne.io.kit.kit import get_kit_info
 from mne_bids import (write_raw_bids, read_raw_bids, BIDSPath,
                       write_anat, make_dataset_description,
                       mark_bad_channels, write_meg_calibration,
-                      write_meg_crosstalk, get_entities_from_fname,
-                      make_minimal_subject_dir)
+                      write_meg_crosstalk, get_entities_from_fname)
 from mne_bids.utils import (_stamp_to_dt, _get_anonymization_daysback,
                             get_anonymization_daysback, _write_json)
 from mne_bids.tsv_handler import _from_tsv, _to_tsv
@@ -1524,15 +1522,6 @@ def _check_anat_json(bids_path):
         coords = anat_dict[point_list[i]]
         np.testing.assert_array_equal(np.asarray(coords, dtype=int),
                                       point)
-
-
-@requires_freesurfer('mri_convert')
-def test_make_minimal_subject_dir(tmpdir):
-    """Test making a minimal subjects directory for coordinate transforms."""
-    data_path = testing.data_path()
-    t1_bids_path = Path(data_path) / 'subjects' / 'sample' / 'mri' / 'T1.mgz'
-    make_minimal_subject_dir(t1_bids_path, tmpdir, 'test')
-    assert op.isfile(op.join(tmpdir, 'test', 'mri', 'T1.mgz'))
 
 
 @requires_nibabel()
